webgpu: don't immediately reacquire swapchain texture;
lovr.graphics.present does gpu_surface_present but then immediately
acquire the next swapchain image with gpu_surface_acquire.  This is good
on Vulkan because it reduces input latency, but on WebGPU it doesn't
make sense because it will immediately re-acquire the swapchain image
for the current frame, since swapchain images are tied to the
requestAnimationFrame.  Then, on the next frame, we won't re-acquire the
swapchain texture because it's already been acquired, and we'll render
to the swapchain image from the previous frame and submit that, which is
invalid.